pyproject.toml
[project]
name = "papercast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=23.12.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-texttospeech>=2.16.0",
    "google-generativeai>=0.3.0",
    "huggingface-hub>=0.20.0",
    "lark>=1.3.0",
    "mutagen>=1.47.0",
    "mypy>=1.8.0",
    "pydantic>=2.5.0",
    "pylint>=3.0.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tenacity>=8.2.0",
]
